requirements.txt
pillow
tensorflow==1.15.0
werkzeug
